fix: load chat history as conversation context for agent
The WebSocket and REST handlers were only sending the current message
to the agent with no prior conversation. Now both load the last 20
messages from chat_messages and prepend them so the LLM can recall
earlier exchanges (e.g. remembering the user's name).

Also includes linter-driven improvements: input validation on routes,
Float32 embeddings, FTS5 query sanitization, and minor cleanups.

+++ src/memory/chat-persistence.ts
@@ -135,7 +135,7 @@ async function generateEmbeddings(
     const chunk = chunks[i]!;
     const result = embeddings[i];
     if (result) {
-      const blob = Buffer.from(new Float64Array(result.embedding).buffer);
+      const blob = Buffer.from(new Float32Array(result.embedding).buffer);
       db.prepare(
         "INSERT OR REPLACE INTO embedding_cache (chunk_id, provider_model, embedding, dimensions) VALUES (?, ?, ?, ?)"
       ).run(chunk.id, pKey, blob, result.dimensions);
